Rename to `no-instanceof-builtins`
