+++ docs/rules/no-instanceof-builtins.md
@@ -73,7 +73,7 @@ The matching strategy:
 - `'strict'` - Matches all built-in constructors.
 
 ```js
-"unicorn/no-instanceof-builtin-object": [
+"unicorn/no-instanceof-builtins": [
 	"error",
 	{
 		"strategy": "strict"
@@ -89,7 +89,7 @@ Default: `[]`
 Specify the constructors that should be validated.
 
 ```js
-"unicorn/no-instanceof-builtin-object": [
+"unicorn/no-instanceof-builtins": [
 	"error",
 	{
 		"include": [
@@ -108,7 +108,7 @@ Default: `[]`
 Specifies the constructors that should be excluded, with this rule taking precedence over others.
 
 ```js
-"unicorn/no-instanceof-builtin-object": [
+"unicorn/no-instanceof-builtins": [
 	"error",
 	{
 		"exclude": [
@@ -127,7 +127,7 @@ Default: `false`
 Specifies using [`Error.isError()`](https://github.com/tc39/proposal-is-error) to determine whether it is an error object.
 
 ```js
-"unicorn/no-instanceof-builtin-object": [
+"unicorn/no-instanceof-builtins": [
 	"error",
 	{
 		"strategy": "strict",
