+++ index.js
@@ -5,7 +5,7 @@ import packageJson from './package.json' with {type: 'json'};
 
 const deprecatedRules = createDeprecatedRules({
 	// {ruleId: ReplacementRuleId | ReplacementRuleId[]}, if no replacement, use `{ruleId: []}`
-	'no-instanceof-array': 'unicorn/no-instanceof-builtin-object',
+	'no-instanceof-array': 'unicorn/no-instanceof-builtins',
 });
 
 const externalRules = {
