+++ test/snapshots/no-instanceof-builtins.js.md
@@ -1,6 +1,6 @@
-# Snapshot report for `test/no-instanceof-builtin-object.js`
+# Snapshot report for `test/no-instanceof-builtins.js`
 
-The actual snapshot is saved in `no-instanceof-builtin-object.js.snap`.
+The actual snapshot is saved in `no-instanceof-builtins.js.snap`.
 
 Generated by [AVA](https://avajs.dev).
